Simplify conditional expressions that are of the form 'expr ? true : false'
