+++ src/EditorFeatures/CSharpTest/UseConditionalExpression/UseConditionalExpressionForReturnTests.cs
@@ -705,5 +705,123 @@ class C
     }
 }");
         }
+
+        [Fact, Trait(Traits.Feature, Traits.Features.CodeActionsUseConditionalExpression)]
+        public async Task TestReturnTrueFalse1()
+        {
+            await TestInRegularAndScriptAsync(
+@"
+class C
+{
+    bool M(int a)
+    {
+        [||]if (a == 0)
+        {
+            return true;
+        }
+        else
+        {
+            return false;
+        }
+    }
+}",
+@"
+class C
+{
+    bool M(int a)
+    {
+        return a == 0;
+    }
+}");
+        }
+
+        [Fact, Trait(Traits.Feature, Traits.Features.CodeActionsUseConditionalExpression)]
+        public async Task TestReturnTrueFalse2()
+        {
+            await TestInRegularAndScriptAsync(
+@"
+class C
+{
+    bool M(int a)
+    {
+        [||]if (a == 0)
+        {
+            return false;
+        }
+        else
+        {
+            return true;
+        }
+    }
+}",
+@"
+class C
+{
+    bool M(int a)
+    {
+        return a != 0;
+    }
+}");
+        }
+
+        [Fact, Trait(Traits.Feature, Traits.Features.CodeActionsUseConditionalExpression)]
+        public async Task TestReturnTrueFalse3()
+        {
+            await TestInRegularAndScriptAsync(
+@"
+class C
+{
+    bool M(int a)
+    {
+        [||]if (a == 0)
+        {
+            return false;
+        }
+
+        return true;
+    }
+}",
+@"
+class C
+{
+    bool M(int a)
+    {
+        return a != 0;
+    }
+}");
+        }
+
+        [Fact, Trait(Traits.Feature, Traits.Features.CodeActionsUseConditionalExpression)]
+        public async Task TestReturnTrueFalse4()
+        {
+            await TestInRegularAndScriptAsync(
+@"
+using System.Collections.Generic;
+
+class C
+{
+    IEnumerable<bool> M(int a)
+    {
+        [||]if (a == 0)
+        {
+            yield return false;
+        }
+        else
+        {
+            yield return true;
+        }
+    }
+}",
+@"
+using System.Collections.Generic;
+
+class C
+{
+    IEnumerable<bool> M(int a)
+    {
+        yield return a != 0;
+    }
+}");
+        }
     }
 }
